Extract the common logic of member access resolution
