+++ src/lang/defs/member.rs
@@ -1,6 +1,10 @@
-use cairo_lang_defs::ids::{MemberId, NamedLanguageElementId};
+use cairo_lang_defs::ids::{FunctionWithBodyId, MemberId, NamedLanguageElementId};
+use cairo_lang_semantic::items::function_with_body::{
+    FunctionWithBodySemantic, SemanticExprLookup,
+};
+use cairo_lang_semantic::{Expr, MemberAccessKind};
 use cairo_lang_syntax::node::ids::SyntaxStablePtrId;
-use cairo_lang_syntax::node::{SyntaxNode, TypedSyntaxNode, ast};
+use cairo_lang_syntax::node::{SyntaxNode, TypedStablePtr, TypedSyntaxNode, ast};
 
 use crate::lang::db::AnalysisDatabase;
 use crate::lang::defs::ItemDef;
@@ -44,4 +48,47 @@ impl<'db> MemberDef<'db> {
     pub fn name(&self, db: &'db AnalysisDatabase) -> String {
         self.member_id.name(db).to_string(db)
     }
+}
+
+/// Resolves the member accessed by a member access expression (e.g. `self.<ident>`) containing
+/// `node`, if `node` is a part of the right-hand side of that expression.
+pub fn resolve_accessed_member<'db>(
+    db: &'db AnalysisDatabase,
+    node: SyntaxNode<'db>,
+    function_with_body: FunctionWithBodyId<'db>,
+) -> Option<MemberId<'db>> {
+    let binary_expr = node.ancestor_of_type::<ast::ExprBinary>(db)?;
+
+    let expr_id =
+        db.lookup_expr_by_ptr(function_with_body, binary_expr.stable_ptr(db).into()).ok()?;
+
+    // Desnap the binary expression to the member access expression.
+    let expr_member_access = match db.expr_semantic(function_with_body, expr_id) {
+        Expr::MemberAccess(expr_member_access) => expr_member_access,
+        Expr::Snapshot(expr_snapshot) => {
+            match db.expr_semantic(function_with_body, expr_snapshot.inner) {
+                Expr::MemberAccess(expr_member_access) => expr_member_access,
+                _ => return None,
+            }
+        }
+        _ => return None,
+    };
+
+    let pointer_to_rhs = binary_expr.rhs(db).stable_ptr(db).untyped();
+
+    let mut current_node = node;
+    // Check if the node points to a member, not a struct variable.
+    while pointer_to_rhs != current_node.stable_ptr(db) {
+        // If we found the node with the binary expression, then we're sure we won't find the
+        // node with the member.
+        if current_node.stable_ptr(db) == binary_expr.stable_ptr(db).untyped() {
+            return None;
+        }
+        current_node = current_node.parent(db)?;
+    }
+
+    let MemberAccessKind::Struct { member_id, .. } = expr_member_access.kind else {
+        return None;
+    };
+    Some(member_id)
 }
